Organize Methods better based on function
